+++ src/core/util/packed/monotonic_block_packed_reader.rs
@@ -11,9 +11,9 @@ use std::sync::Arc;
 
 /// Provides random access to a stream written with MonotonicBlockPackedWriter
 pub struct MonotonicBlockPackedReader {
-    block_shift: u32,
-    block_mask: u32,
-    value_count: u64,
+    block_shift: usize,
+    block_mask: usize,
+    value_count: usize,
     min_values: Vec<i64>,
     averages: Vec<f32>,
     sub_readers: Vec<Box<packed_misc::Reader>>,
@@ -32,7 +32,7 @@ impl MonotonicBlockPackedReader {
         input: &mut IndexInput,
         packed_ints_version: i32,
         block_size: usize,
-        value_count: u64,
+        value_count: usize,
         direct: bool,
     ) -> Result<MonotonicBlockPackedReader> {
         let block_shift = packed_misc::check_block_size(
@@ -40,7 +40,7 @@ impl MonotonicBlockPackedReader {
             packed_misc::MIN_BLOCK_SIZE,
             packed_misc::MAX_BLOCK_SIZE,
         );
-        let block_mask = (block_size - 1) as u32;
+        let block_mask = block_size - 1;
         let num_blocks = packed_misc::num_blocks(value_count, block_size);
         let mut min_values = vec![0_i64; num_blocks];
         let mut averages = vec![0.0_f32; num_blocks];
@@ -58,8 +58,8 @@ impl MonotonicBlockPackedReader {
             if bits_per_value == 0 {
                 sub_readers.push(Box::new(PackedIntsNullReader::new(block_size)));
             } else {
-                let left = value_count - (i as u64) * (block_size as u64);
-                let size = ::std::cmp::min(left, block_size as u64) as u32;
+                let left = value_count - i * block_size;
+                let size = ::std::cmp::min(left, block_size);
                 if direct {
                     unimplemented!();
                 } else {
@@ -67,7 +67,7 @@ impl MonotonicBlockPackedReader {
                         input,
                         packed_misc::Format::Packed,
                         packed_ints_version,
-                        size as usize,
+                        size,
                         bits_per_value,
                     )?;
                     sub_readers.push(one_reader);
@@ -86,7 +86,7 @@ impl MonotonicBlockPackedReader {
     }
 
     /// Returns the number of values
-    pub fn size(&self) -> u64 {
+    pub fn size(&self) -> usize {
         self.value_count
     }
 }
@@ -100,8 +100,8 @@ impl LongValues for MonotonicBlockPackedReader {
         if !(index >= 0 && index < self.value_count as i64) {
             bail!(IllegalArgument(format!("index {} out of range", index)))
         }
-        let block = (index >> (self.block_shift as usize)) as usize;
-        let idx = (index & (i64::from(self.block_mask))) as i32;
+        let block = (index >> self.block_shift) as usize;
+        let idx = (index & (self.block_mask as i64)) as i32;
         let val = Self::expected(self.min_values[block], self.averages[block], idx)
             + self.sub_readers[block].get(idx as usize);
         Ok((val, ctx))
